Create table component for use in other apps.
