+++ src/App.css
@@ -17,10 +17,6 @@ body {
   margin: 0;
 }
 
-table {
-  border-collapse: collapse;
-}
-
 fieldset {
   margin: 0;
   border: none;
